refactor(docs): update plugin examples
