+++ packages/site/examples/plugin/bubble-sets/index.en.md
@@ -1,4 +1,3 @@
 ---
 title: Bubblesets
-order: 1
 ---
