+++ packages/site/examples/plugin/bubble-sets/index.zh.md
@@ -1,4 +1,3 @@
 ---
 title: Bubblesets 气泡集
-order: 1
 ---
