+++ packages/site/examples/plugin/contextMenu/index.en.md
@@ -1,27 +1,3 @@
 ---
 title: ContextMenu
-order: 2
 ---
-
-Using contextMenu in G6.
-
-## Usage
-
-The demo below show how to use contextMenu on a graph. Menu's style can be defined by the CSS:
-
-```css
-/* Define the CSS with the id of your menu */
-  #contextMenu {
-    position: absolute;
-    /* ... Other styles */
-  }
-  #contextMenu li {
-    cursor: pointer;
-		/* ... Other styles */
-  }
-  #contextMenu li:hover {
-    color: #aaa;
-    /* ... Other styles */
-  }
-`);
-```
