+++ packages/site/examples/plugin/grid-line/demo/basic.js
@@ -1,0 +1,38 @@
+import { Graph } from '@antv/g6';
+
+const data = {
+  nodes: [{ id: 'node-0' }, { id: 'node-1' }, { id: 'node-2' }, { id: 'node-3' }, { id: 'node-4' }, { id: 'node-5' }],
+  edges: [
+    { source: 'node-0', target: 'node-1' },
+    { source: 'node-0', target: 'node-2' },
+    { source: 'node-0', target: 'node-3' },
+    { source: 'node-0', target: 'node-4' },
+    { source: 'node-1', target: 'node-0' },
+    { source: 'node-2', target: 'node-0' },
+    { source: 'node-3', target: 'node-0' },
+    { source: 'node-4', target: 'node-0' },
+    { source: 'node-5', target: 'node-0' },
+  ],
+};
+
+const graph = new Graph({
+  container: 'container',
+  data,
+  layout: { type: 'grid' },
+  behaviors: ['drag-canvas'],
+  plugins: [{ key: 'grid-line', type: 'grid-line', follow: false }],
+});
+
+graph.render();
+
+window.addPanel((gui) => {
+  gui
+    .add({ follow: false }, 'follow')
+    .name('Follow')
+    .onChange((value) => {
+      graph.updatePlugin({
+        key: 'grid-line',
+        follow: value,
+      });
+    });
+});
