+++ packages/site/examples/plugin/grid-line/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "default.ts",
+      "filename": "basic.js",
       "title": {
         "zh": "画布网格",
         "en": "Grid Line"
