+++ packages/site/examples/plugin/grid-line/index.en.md
@@ -1,8 +1,3 @@
 ---
 title: Grid Line
-order: 3
 ---
-
-## Usage
-
-The demo below show how to use Grid Line Plugin on a graph.
