+++ packages/site/examples/plugin/grid-line/index.zh.md
@@ -1,8 +1,3 @@
 ---
 title: Grid Line 网格线
-order: 3
 ---
-
-## 使用指南
-
-下面的代码演示展示了如何在图上使用网格线。
