+++ packages/site/examples/plugin/history/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "default.ts",
+      "filename": "basic.js",
       "title": {
         "zh": "撤销重做",
         "en": "Redo/Undo"
