+++ packages/site/examples/plugin/history/index.en.md
@@ -1,4 +1,3 @@
 ---
 title: History
-order: 4
 ---
