+++ packages/site/examples/plugin/history/index.zh.md
@@ -1,4 +1,3 @@
 ---
 title: History 撤销重做
-order: 4
 ---
