+++ packages/site/examples/plugin/hull/index.en.md
@@ -1,6 +1,3 @@
 ---
 title: Hull
-order: 5
 ---
-
-Use a smooth contour to wrap a specified set of nodes.
