+++ packages/site/examples/plugin/hull/index.zh.md
@@ -1,6 +1,3 @@
 ---
 title: Hull 节点集轮廓包裹
-order: 5
 ---
-
-使用平滑的轮廓来包裹指定的一组节点集合。
