+++ packages/site/examples/plugin/legend/demo/style.js
