+++ packages/site/examples/plugin/legend/index.en.md
@@ -1,8 +1,3 @@
 ---
 title: Legend
-order: 6
 ---
-
-## Usage
-
-The demo below show how to use Legend Plugin on a graph.
