+++ packages/site/examples/plugin/legend/index.zh.md
@@ -1,8 +1,3 @@
 ---
 title: Legend 图例
-order: 6
 ---
-
-## 使用指南
-
-下面的代码演示展示了如何在图上使用 Legend。
