--- packages/site/examples/plugin/overview/demo/default.ts
@@ -1,50 +1,0 @@
-import { Graph as BaseGraph, Extensions, Util, extend } from '@antv/g6';
-
-const Graph = extend(BaseGraph, {
-  plugins: {
-    toolbar: Extensions.Toolbar,
-    minimap: Extensions.Minimap,
-    grid: Extensions.Grid,
-  },
-  behaviors: {
-    'activate-relations': Extensions.ActivateRelations,
-  },
-});
-
-import insertCss from 'insert-css';
-const container = document.getElementById('container') as HTMLElement;
-const width = container.scrollWidth;
-const height = (container.scrollHeight || 500) - 110;
-const data = Util.mock(6).circle();
-
-new Graph({
-  container,
-  width,
-  height,
-  data,
-  modes: {
-    default: ['brush-select', 'zoom-canvas', 'activate-relations', 'drag-canvas', 'drag-element'],
-  },
-  plugins: [
-    'toolbar',
-    'minimap',
-    {
-      key: 'default-grid',
-      type: 'grid',
-      img: 'url(data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDAiIGhlaWdodD0iNDAiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PGRlZnM+PHBhdHRlcm4gaWQ9ImdyaWQiIHdpZHRoPSI0MCIgaGVpZ2h0PSI0MCIgcGF0dGVyblVuaXRzPSJ1c2VyU3BhY2VPblVzZSI+PHBhdGggZD0iTSAwIDEwIEwgNDAgMTAgTSAxMCAwIEwgMTAgNDAgTSAwIDIwIEwgNDAgMjAgTSAyMCAwIEwgMjAgNDAgTSAwIDMwIEwgNDAgMzAgTSAzMCAwIEwgMzAgNDAiIGZpbGw9Im5vbmUiIHN0cm9rZT0iI2UwZTBlMCIgb3BhY2l0eT0iMC4yIiBzdHJva2Utd2lkdGg9IjEiLz48cGF0aCBkPSJNIDQwIDAgTCAwIDAgMCA0MCIgZmlsbD0ibm9uZSIgc3Ryb2tlPSIjZTBlMGUwIiBzdHJva2Utd2lkdGg9IjEiLz48L3BhdHRlcm4+PC9kZWZzPjxyZWN0IHdpZHRoPSIxMDAlIiBoZWlnaHQ9IjEwMCUiIGZpbGw9InVybCgjZ3JpZCkiLz48L3N2Zz4=)',
-    },
-  ],
-});
-
-/** set the style of minimap  */
-insertCss(`
-  .g6-minimap {
-    position:absolute;
-    bottom:40px;
-    right:40px;
-    box-shadow:0 2px 2px 0 rgba(0,0,0,0.14), 0 3px 1px -2px rgba(0,0,0,0.12), 0 1px 5px 0 rgba(0,0,0,0.2);
-  }
-  .g6-minimap-viewport {
-    border: 2px solid rgb(25, 128, 255);
-  }
-`);
