--- packages/site/examples/plugin/overview/demo/meta.json
@@ -1,16 +1,0 @@
-{
-  "title": {
-    "zh": "中文分类",
-    "en": "Category"
-  },
-  "demos": [
-    {
-      "filename": "default.ts",
-      "title": {
-        "zh": "总览",
-        "en": "Overview"
-      },
-      "screenshot": "https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*JtOPS4unj58AAAAAAAAAAAAADmJ7AQ/original"
-    }
-  ]
-}
