--- packages/site/examples/plugin/overview/index.en.md
@@ -1,6 +1,0 @@
----
-title: Overview
-order: 0
----
-
-Overview all plugins built-in G6
