--- packages/site/examples/plugin/overview/index.zh.md
@@ -1,6 +1,0 @@
----
-title: 插件总览
-order: 0
----
-
-总揽 G6 中的插件
