+++ packages/site/examples/plugin/timebar/demo/chart.js
