+++ packages/site/examples/plugin/timebar/demo/time.js
