+++ packages/site/examples/plugin/timebar/index.en.md
@@ -1,4 +1,3 @@
 ---
 title: Timebar
-order: 7
 ---
