+++ packages/site/examples/plugin/timebar/index.zh.md
@@ -1,4 +1,3 @@
 ---
 title: Timebar 时间条
-order: 7
 ---
