+++ packages/site/examples/plugin/toolbar/demo/custom.js
@@ -1,5 +1,10 @@
 import { Graph } from '@antv/g6';
 
+// Use your own iconfont.
+const iconFont = document.createElement('script');
+iconFont.src = '//at.alicdn.com/t/font_8d5l8fzk5b87iudi.js';
+document.head.appendChild(iconFont);
+
 const data = {
   nodes: [{ id: 'node-0' }, { id: 'node-1' }, { id: 'node-2' }, { id: 'node-3' }, { id: 'node-4' }, { id: 'node-5' }],
   edges: [
@@ -18,9 +23,26 @@ const data = {
 const graph = new Graph({
   container: 'container',
   data,
-  layout: { type: 'grid' },
-  behaviors: ['drag-canvas'],
-  plugins: [{ type: 'grid-line', follow: false }],
+  layout: {
+    type: 'grid',
+  },
+  behaviors: ['zoom-canvas', 'drag-canvas', 'drag-element'],
+  plugins: [
+    {
+      type: 'toolbar',
+      position: 'right-top',
+      onClick: (item) => {
+        alert('item clicked:' + item);
+      },
+      getItems: () => {
+        return [
+          { id: 'icon-xinjian', value: 'new' },
+          { id: 'icon-fenxiang', value: 'share' },
+          { id: 'icon-chexiao', value: 'undo' },
+        ];
+      },
+    },
+  ],
 });
 
 graph.render();
