+++ packages/site/examples/plugin/toolbar/index.en.md
@@ -1,16 +1,3 @@
 ---
 title: ToolBar
-order: 8
 ---
-
-ToolBar is a build-in Component in G6, but not reigstered by default. It needs to be imported into the code and registered with the `extend` method, and then you could configured it to the graph instance.
-
-## Usage
-
-The demo below show how to use toolbar on graph. Toolbar's style can be defined by the CSS with class name `g6-component-toolbar`:
-
-```css
-.g6-component-toolbar {
-  // css styles
-}
-```
