+++ packages/site/examples/plugin/toolbar/index.zh.md
@@ -1,14 +1,3 @@
 ---
 title: Toolbar 工具栏
-order: 8
 ---
-
-## 使用指南
-
-下面的代码演示展示了如何在图上使用 ToolBar。ToolBar 的样式可以使用 G6 内置的，也可以完全自定义 ToolBar 的内容，要修改内置 ToolBar 的样式，只需要修改 g6-component-toolbar 的样式：
-
-```css
-.g6-component-toolbar {
-  // 自定义 CSS 内容
-}
-```
