+++ packages/site/examples/plugin/tooltip/index.en.md
@@ -1,4 +1,3 @@
 ---
 title: Tooltip
-order: 9
 ---
