+++ packages/site/examples/plugin/tooltip/index.zh.md
@@ -1,4 +1,3 @@
 ---
 title: Tooltip 提示框
-order: 9
 ---
