+++ packages/site/examples/plugin/watermark/demo/repeat.js
