+++ packages/site/examples/plugin/watermark/demo/text.js
