+++ packages/site/examples/plugin/watermark/index.en.md
@@ -1,4 +1,3 @@
 ---
 title: Watermark
-order: 10
 ---
