+++ packages/site/examples/plugin/watermark/index.zh.md
@@ -1,4 +1,3 @@
 ---
 title: Watermark 水印
-order: 10
 ---
